Move getAzureAppServiceClient to AbstarctAzureAppServiceMojo
