+++ azure-functions-maven-plugin/src/main/java/com/microsoft/azure/maven/function/AbstractFunctionMojo.java
@@ -6,26 +6,13 @@
 package com.microsoft.azure.maven.function;
 
 import com.microsoft.azure.maven.AbstractAppServiceMojo;
-import com.microsoft.azure.maven.auth.AzureAuthFailureException;
-import com.microsoft.azure.toolkit.lib.Azure;
-import com.microsoft.azure.toolkit.lib.appservice.AzureAppService;
 import com.microsoft.azure.toolkit.lib.appservice.service.IFunctionApp;
-import com.microsoft.azure.toolkit.lib.auth.Account;
-import com.microsoft.azure.toolkit.lib.auth.AzureAccount;
-import com.microsoft.azure.toolkit.lib.auth.exception.AzureLoginException;
-import com.microsoft.azure.toolkit.lib.common.exception.AzureExecutionException;
 import com.microsoft.azure.toolkit.lib.common.exception.AzureToolkitRuntimeException;
-import com.microsoft.azure.toolkit.lib.common.logging.Log;
-import com.microsoft.azure.toolkit.lib.common.model.Subscription;
-import com.microsoft.azure.toolkit.lib.common.utils.TextUtils;
 import com.microsoft.azure.toolkit.lib.legacy.function.configurations.RuntimeConfiguration;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.plugins.annotations.Parameter;
 
 import java.io.File;
-import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
 public abstract class AbstractFunctionMojo extends AbstractAppServiceMojo {
@@ -121,37 +108,7 @@ public abstract class AbstractFunctionMojo extends AbstractAppServiceMojo {
         return disableAppInsights;
     }
 
-    private AzureAppService appServiceClient;
-
-    protected AzureAppService getOrCreateAzureAppServiceClient() throws AzureExecutionException {
-        if (appServiceClient == null) {
-            try {
-                final Account account = getAzureAccount();
-                final List<Subscription> subscriptions = account.getSubscriptions();
-                final String targetSubscriptionId = getTargetSubscriptionId(getSubscriptionId(), subscriptions, account.getSelectedSubscriptions());
-                checkSubscription(subscriptions, targetSubscriptionId);
-                com.microsoft.azure.toolkit.lib.Azure.az(AzureAccount.class).account().selectSubscription(Collections.singletonList(targetSubscriptionId));
-                appServiceClient = Azure.az(AzureAppService.class).subscription(targetSubscriptionId);
-                printCurrentSubscription(appServiceClient);
-            } catch (AzureLoginException | AzureExecutionException | IOException e) {
-                throw new AzureExecutionException(String.format("Cannot authenticate due to error %s", e.getMessage()), e);
-            }
-        }
-        return appServiceClient;
-    }
-
-    // todo: Replace same method in AbstractAzureMojo after function track2 migration
-    protected void printCurrentSubscription(AzureAppService appServiceClient) {
-        if (appServiceClient == null) {
-            return;
-        }
-        final Subscription subscription = appServiceClient.getDefaultSubscription();
-        if (subscription != null) {
-            Log.info(String.format(SUBSCRIPTION_TEMPLATE, TextUtils.cyan(subscription.getName()), TextUtils.cyan(subscription.getId())));
-        }
-    }
-
-    public IFunctionApp getFunctionApp() throws AzureAuthFailureException, AzureExecutionException {
+    public IFunctionApp getFunctionApp() {
         return getOrCreateAzureAppServiceClient().functionApp(getResourceGroup(), getAppName());
     }
